backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
passlib[bcrypt]>=1.7.4
pydantic>=2.5.0
python-dotenv>=1.0.0
sqlmodel>=0.0.16
pydantic-settings>=2.0.3
SQLAlchemy>=2.0.23
aiosqlite>=0.19.0
asyncpg>=0.29.0